add kxbmap-configs & make separate spaces for scalaz 7.1/7.2
this reflects a real split in the community.

separate 7.2 space is needed in order to add kxbmap-configs.
(a library we don't absolutely need to have, but we want it because
it caught a couple of 2.12 regressions. and in any case, we may
need the scalaz 7.2 space for other things in the future.)